% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab.R
\name{as.index}
\alias{as.index}
\title{as.index: equivalent of MATLAB's ind2sub}
\usage{
as.index(rows, rownum, colnum)
}
\arguments{
\item{rows}{Total number of rows in the matrix. Scalar value.}

\item{rownum}{row subscript. vector of any length.}

\item{colnum}{column subscript. vector of any length.}
}
\value{
A numeric vector of linear indices
}
\description{
Given the row  and column  of elements in an m × n matrix,
compute the single index which can be used to access those
elements of A (also works if r and c are vectors)
}
