% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{betaABfromMeanKappa}
\alias{betaABfromMeanKappa}
\title{betaABfromMeanKappa
Beta distribution shape parameters recovered from mean and concentration:}
\usage{
betaABfromMeanKappa(mean, kappa)
}
\arguments{
\item{mean}{numeric scalar in range (0,1)}

\item{kappa}{numeric scalar in range (0,Inf)}
}
\value{
Named list with 2 elements:
a = Beta shape parameter #1,
b = Beta shape parameter #2
}
\description{
betaABfromMeanKappa
Beta distribution shape parameters recovered from mean and concentration:
}
\examples{
betaABfromMeanKappa(mean = .6, kappa=15)
}
\author{
John Krushke, in Kruschke, J. K. (2015). Doing Bayesian Data Analysis, Second Edition:
A Tutorial with R, JAGS, and Stan. Academic Press / Elsevier.
}

