% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{betaABfromMeanSD}
\alias{betaABfromMeanSD}
\title{betaABfromMeanSD
Beta distribution shape parameters recovered from mean and standard deviation}
\usage{
betaABfromMeanSD(mean, sd)
}
\arguments{
\item{mean}{numeric scalar in range (0,1)}

\item{sdnumeric}{scalar in range (0,Inf)}
}
\value{
Named list with 2 elements:
a = Beta shape parameter #1,
b = Beta shape parameter #2
}
\description{
betaABfromMeanSD
Beta distribution shape parameters recovered from mean and standard deviation
}
\examples{
betaABfromMeanSD(mean =.5, sd = 1)
}
\author{
John Krushke, in Kruschke, J. K. (2015). Doing Bayesian Data Analysis, Second Edition:
A Tutorial with R, JAGS, and Stan. Academic Press / Elsevier.
}

