% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{betaABfromModeKappa}
\alias{betaABfromModeKappa}
\title{betaABfromModeKappa
Beta distribution shape parameters recovered from mode and concentration:}
\usage{
betaABfromModeKappa(mode, kappa)
}
\arguments{
\item{mode}{numeric scalar in range (0,1)}

\item{kappa}{numeric scalar in range (0,Inf)}
}
\value{
Named list with 2 elements:
a = Beta shape parameter #1,
b = Beta shape parameter #2
}
\description{
betaABfromModeKappa
Beta distribution shape parameters recovered from mode and concentration:
}
\examples{
betaABfromModeKappa(mode = .4, kappa=10)
}
\author{
John Krushke, in Kruschke, J. K. (2015). Doing Bayesian Data Analysis, Second Edition:
A Tutorial with R, JAGS, and Stan. Academic Press / Elsevier.
}

