% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorStats.R
\name{binomialg2}
\alias{binomialg2}
\title{binomial g2}
\usage{
binomialg2(obs, pred, N)
}
\arguments{
\item{obs}{Observed Proportions}

\item{pred}{Predicted Proportions}

\item{N}{Cell Frequencies}
}
\value{
Scalar numeric vector
}
\description{
Likelihood ratio statistic for binomial random variable
}

