% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorStats.R
\name{binomialLL}
\alias{binomialLL}
\title{Binomial Log Likelihood}
\usage{
binomialLL(obs, pred, N)
}
\arguments{
\item{obs}{Observed Proportions}

\item{pred}{Predicted Proportions}

\item{N}{Cell Frequencies}
}
\value{
Scalar numeric vector
}
\description{
Binomial Log Likelihood
}

