% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{estBetaParams}
\alias{estBetaParams}
\title{estBetaParams
Solve for the shape parameters of the beta distribution given its mean and
variance}
\usage{
estBetaParams(mu, var)
}
\arguments{
\item{mu}{A numeric vector of length 1, [0,1]}

\item{var}{A numeric vector of length 1, [0,1]}
}
\value{
A list with 2 elements, alpha and beta
}
\description{
estBetaParams
Solve for the shape parameters of the beta distribution given its mean and
variance
}
\examples{
estBetaParams(mu = .5, var = .25)
}

