% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{gammaParams}
\alias{gammaParams}
\title{Gamma Shape and Rate
Determine the shape and rate parameter of a gamma distribution, given its central
tendency and spread.See the 'Details' section for constraints on which arguments
may be used together.}
\usage{
gammaParams(mean, mode, sd)
}
\arguments{
\item{mean}{A numeric vector}

\item{mode}{A numeric vector}

\item{sd}{A numeric vector}
}
\value{
Named list with 2 elements:
S = Gamma shape parameter (alpha)
R = Gamma rate parameter (beta)
}
\description{
Gamma Shape and Rate
Determine the shape and rate parameter of a gamma distribution, given its central
tendency and spread.See the 'Details' section for constraints on which arguments
may be used together.
}
\details{
Only two parameters may be specified together. Specifically, one parameter
relating to the distributions central tendency (i.e. mean or mode) must be specified with
the standard deviation.
}
\examples{
gammaParams(mean = .1.5, sd = 3)
gammaParams(mode = 2, sd = 3)

# Throws an error, because mean and mode cannot be used to find the shape and rate.
\dontrun{gammaParams(mean = 2, mode = 3)}
}
\author{
Will Hopper (2016)
Inspired by functions writen by John Krushke to accompany his "Doing Bayesian Data Analysis" book.
Kruschke, J. K. (2015). Doing Bayesian Data Analysis, Second Edition:
A Tutorial with R, JAGS, and Stan. Academic Press / Elsevier.
}

