% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gammaShRaFromModeSD}
\alias{gammaShRaFromModeSD}
\title{gammaShRaFromModeSD
Gamma distribution shape and rate parameters recovered fromm mode and sd}
\usage{
gammaShRaFromModeSD(mode, sd)
}
\arguments{
\item{mode}{a numeric vector}

\item{sd}{a numeric vector}
}
\value{
A named list with 2 elements:
shape = Gamma distribution shape paratmeter
rate = Gamma distribution rate parameter
}
\description{
gammaShRaFromModeSD
Gamma distribution shape and rate parameters recovered fromm mode and sd
}
\examples{
gammaShRaFromModeSD(mode = 2, sd = 3)
}
\author{
John Krushke, in Kruschke, J. K. (2015). Doing Bayesian Data Analysis, Second Edition:
A Tutorial with R, JAGS, and Stan. Academic Press / Elsevier.
}

