% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{getRandString}
\alias{getRandString}
\title{getRandString}
\usage{
getRandString(len = 12)
}
\arguments{
\item{len}{The number of characters in the string created}
}
\value{
A character vector of length 1
}
\description{
Create a psuedo-random alphanumeric string
}
\details{
The characters in the string are sampled with a replacement from a
 pool of the numbers 0 through 9, A-Z, and a-z, with equal sampling probability for all
 items in the pool.

 The string returned is not gauranteed to be unique across runs, so take care when using this function
 to create unique identifiers (e.g., for files names). Consider packages such as
 \href{https://cran.r-project.org/web/packages/ids/}{ids} or \href{https://cran.r-project.org/web/packages/uuid}{uuid} in situations where all generated identifies
  must be globally unique.
}
\examples{
getRandString(len = 12)

}
