% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarising.R
\name{heirarchicalSummary}
\alias{heirarchicalSummary}
\title{Hierarchical Summary}
\usage{
heirarchicalSummary(collapse, hold, rawData, ...)
}
\arguments{
\item{collapse}{A character vector of variables to group the data frame by,
which will be succesively collapsed over.}

\item{hold}{A character vector of variables to consistently group the data frame by,
even while other variables are collapsed}

\item{rawData}{The data frame to summmarise iteratively}

\item{...}{Name-value pairs of summary functions like mean(), sd() etc.
See details for warnings about naming conventions}
}
\value{
A list of tbl_df's, the same length as the character vector passed as the
 \code{collapse} argument
}
\description{
Summarise a data frame at cascading levels of granularity
}
\examples{
derp <- heirarchicalSummary(collapse = c("subject","group"),
                            hold = c("practice","other_type",
                                     "prac_score", "other_prac_acc"),
                            rawData=ungroup(cbind(LB4L_allSs,n=1)),
                            weighted.mean(final_score,n), mean(final_score),
                            n = sum(n))
str(derp)
}

