% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_helpers.R
\name{load_knitr_cache}
\alias{load_knitr_cache}
\title{Load Knitr Cache}
\usage{
load_knitr_cache(path, envir = parent.frame())
}
\arguments{
\item{path}{The path to the knitr cache directory.}

\item{envir}{The environment to load the cached objects and packges into.
Defaults to the current calling environment}
}
\value{
Returns `NULL` invisibly
}
\description{
Loads the packages and objects stored in a knitr cache. Useful for restoring the
internal state of a knitted document for use in an interactive session without forcing you to
re-run all the code interactively.
}
\details{
The knitr cache directory is usually stored in the same directory as the document. The knitr cache directory naming schema is
`$DOCUMENTNAME$_cache/$OUTPUTFORMAT$/`. So, if your document named "Analysis.Rmd" was knit to HTML
and cached, then the cache would be stored at `Analysis_cache/html/`.

The cached objects are loaded as promises.
}
\examples{
\dontrun{
load_knitr_cache("Analysis_cache/html")
}
}
