% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorStats.R
\name{multinomialg2}
\alias{multinomialg2}
\title{Multinomial g2}
\usage{
multinomialg2(obs, pred, N)
}
\arguments{
\item{obs}{Observed Proportions}

\item{pred}{Predicted Proportions}

\item{N}{Cell Frequencies}
}
\value{
Scalar numeric vector
}
\description{
Likelihood ratio statistic for multinomial random variable
}

