% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorStats.R
\name{multinomialLL}
\alias{multinomialLL}
\title{Multinomial Log Likelihood}
\usage{
multinomialLL(obs, pred, N)
}
\arguments{
\item{obs}{Observed Proportions}

\item{pred}{Predicted Proportions}

\item{N}{Cell Frequencies}
}
\value{
Scalar numeric vector
}
\description{
Multinomial Log Likelihood
}
