% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarising.R
\name{performanceBins}
\alias{performanceBins}
\title{Performance Bins}
\usage{
performanceBins(data, bin_by, cutpoints = c(0, 0.25, 0.5, 0.75, 1), ...)
}
\arguments{
\item{data}{data frame to work with}

\item{bin_by}{Character vector defining the variables to group the data by}

\item{cutpoints}{numeric vector of value to bin the measured variable at. Values not within the range
of this vector will be grouped into the NA category. By default, data is binned
into quartiles.}

\item{...}{Name-value pairs of summary functions like mean(), sd() etc.}
}
\value{
An ungrouped table df summmarising the observed count and frequency in each bin
}
\description{
Summarise a variables by groups and bin the observations
}
\examples{
filter(LB4L_allSs, list==1) \%>\%
performanceBins(bin_by = "subject",
                cutpoints=c(0,.25, .75,1),
                acc=mean(final_score))
}

