% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizingOperations.R
\name{prettifyTable}
\alias{prettifyTable}
\title{prettifyTable}
\usage{
prettifyTable(data, rows)
}
\arguments{
\item{data}{The data frame to subset and print}

\item{rows}{The index vector definining the number of rows to print from the first row}
}
\value{
The subset data frame, with an addition row of "...", and all variables
coerced to characters
}
\description{
Pretty-print a truncated data frame or matrix used in markdown documents with "..."
as the last value in each column to indicate there are more rows not shown.
}
\examples{
prettifyTable(Orange, 9)
}
