% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarising.R
\name{sem}
\alias{sem}
\title{Standard Error of the Mean}
\usage{
sem(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric or logical atomic vector}

\item{na.rm}{a logical value indicating whether NA values should be removed from the input.}
}
\value{
A scalar numeric vector
}
\description{
Calculates the standard error of the mean statistic, an estimate of the variability
 of the sampling distribution of the mean. See "details" for equation.
}
\details{
Uses the following forumula:

\deqn{S.E.M. = \sqrt{\frac{Var(x)}{N}}}{S.E.M. =  sqrt(Var(x)/N)}
}
\examples{

x <- runif(30, 5, 2)
sem(x)

}

