% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{SRfromModeSD}
\alias{SRfromModeSD}
\title{SRfromModeSD
Gamma distribution shape and rate parameters recovered from mode and sd}
\usage{
SRfromModeSD(mode, sd)
}
\arguments{
\item{mode}{a numeric vector}

\item{sd}{a numeric vector}
}
\value{
A named list with 2 elements:
shape = Gamma distribution shape paratmeter
rate = Gamma distribution rate parameter
}
\description{
SRfromModeSD
Gamma distribution shape and rate parameters recovered from mode and sd
}
\examples{
\dontrun{SRfromModeSD(mode = 2, sd = 3)}
}
\author{
John Krushke, in Kruschke, J. K. (2015). Doing Bayesian Data Analysis, Second Edition:
A Tutorial with R, JAGS, and Stan. Academic Press / Elsevier.
}

