% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalizingOperations.R
\name{strSort}
\alias{strSort}
\title{String Character Sorting}
\usage{
strSort(x, splitter = ".")
}
\arguments{
\item{x}{A character vector. Factors or numerics will be coerced with `as.character`}

\item{splitter}{}
}
\value{
A character vector
}
\description{
a character vector on a delimeter, and sort the resulting pieces
Useful for sorting the output of `interaction()`.
}
\examples{
x <- c("zZaAG", 'AtC_Z')
strSort(x,splitter="")
}

