% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{t_summary}
\alias{t_summary}
\title{t test summary
performs a t-test comparing the data from two specified groups,
calculates Cohen's d as a measure of effect size, and returns the
results in format suitable for including in a matrix, data frame or table}
\usage{
t_summary(pair, formula, data, ...)
}
\arguments{
\item{pair}{A vector of length two specifying which groups should be compared.}

\item{formula}{A formula identifying the columns holding the dependent and independent
variables in the data frame/matrix.}

\item{data}{A data frame or matrix containing the variables in the formula.}

\item{...}{Other arguments passed to the t.test function}
}
\value{
A named numeric vector, with an element for the Student's t static,
the degrees of freedom used for the t.test, the upper and lower bounds of the confidence
interval, Cohen's, and the p value.
}
\description{
t test summary
performs a t-test comparing the data from two specified groups,
calculates Cohen's d as a measure of effect size, and returns the
results in format suitable for including in a matrix, data frame or table
}
\author{
Will Hopper
}

