% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{varBeta}
\alias{varBeta}
\title{varBeta
Calculate the variance of a beta distribution given its two shape parameters}
\usage{
varBeta(alpha, beta)
}
\arguments{
\item{alpha}{Beta distribution shape parameter #1 [0, Inf]}

\item{beta}{Beta distribution shape parameter #2 [0, Inf]}
}
\value{
A numeric scalar
}
\description{
varBeta
Calculate the variance of a beta distribution given its two shape parameters
}
\examples{
varBeta(alpha=.01, beta = 1)
}

