% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarising.R
\name{WISEsummary}
\alias{WISEsummary}
\title{Title}
\usage{
WISEsummary(data, DV, betweenvars = NULL, withinvars = NULL, idvar = NULL,
  CI_width = 0.95)
}
\arguments{
\item{data}{A data frame}

\item{DV}{Character vector giving the dependent variable}

\item{betweenvars}{Character vector giving the between subject variables}

\item{withinvars}{Character vector giving the within subject variables}

\item{idvar}{Character vector giving the name of the column holding subject
identifiers}

\item{CI_width}{Numeric vector giving the confidence level for computing the
confidence interval boundaries. Must be between 0 and 1, non-inclusive.}
}
\value{
A data frame
}
\description{
Title
}
\examples{
library(tidyr)
data <- data.frame(subject = 1:12,
Round_Mono = c(41L, 57L, 52L, 49L, 47L, 37L, 47L, 41L, 48L, 37L, 32L, 47L),
Square_Mono = c(40L, 56L, 53L, 47L, 48L, 34L, 50L, 40L, 47L, 35L, 31L, 42L),
Round_Color = c(41L, 56L, 53L, 47L, 48L, 35L, 47L, 38L, 49L, 36L, 31L, 42L),
Square_Color = c(37L, 53L, 50L, 47L, 47L, 36L, 46L, 40L, 45L, 35L, 33L, 42L)) \%>\%
  gather(condition, time, Round_Mono:Square_Color) \%>\%
  separate(condition, c("Shape","Color"), sep = "_") \%>\%
  setNames(tolower(names(.)))

data_summary <- WISEsummary(data, DV = "time", withinvars = c("shape","color"),
                            idvar = "subject")

}

