% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdt.R
\name{dprime}
\alias{dprime}
\title{d'
Calculate the d-prime statistic using the hit and false alarm rate
to stimuli from the target and lure distribution, respectively.}
\usage{
dprime(targets, lures, correct = FALSE)
}
\arguments{
\item{correct}{If true, the 1/2n correction is applied for all yes or all no responses so that
the dprime statistic is not Inf or NaN after the z-transform is applied to the Hit and False Alarm Rates}

\item{HR}{A logical vector of yes (TRUE)/no (FALSE) responses to stimili from
the target (S1) distribution. Used to calculate the Hit Rate.}

\item{FA}{logical vector of yes (TRUE)/no (FALSE) responses to stimili from
the lure (S0) distribution. Used to calculate the False Alarm Rate.}
}
\value{
A scalar numeric vector
}
\description{
d'
Calculate the d-prime statistic using the hit and false alarm rate
to stimuli from the target and lure distribution, respectively.
}
\examples{
dprime(c(0,0,1,1), c(1,0,0,0))
dprime(c(0,0,0,0), c(0,0,0,0), correct = TRUE)
dprime(c(0,0,0,0), c(0,0,0,0)) # NaN
}

