% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{getRandString}
\alias{getRandString}
\title{getRandString}
\usage{
getRandString(len = 12)
}
\arguments{
\item{len}{}
}
\value{
A character vector of length 1
}
\description{
getRandString
}
\examples{
getRandString(len = 12)
}
