% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{meanBeta}
\alias{meanBeta}
\title{meanBeta
Calculate the mean of a beta distribution given its two shape parameters}
\usage{
meanBeta(alpha, beta)
}
\arguments{
\item{alpha}{Beta distribution shape parameter #1 [0, Inf]}

\item{beta}{Beta distribution shape parameter #2 [0, Inf]}
}
\value{
A numeric scalar
}
\description{
meanBeta
Calculate the mean of a beta distribution given its two shape parameters
}
\examples{
meanBeta(alpha=.01, beta = 1)
}

