% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarising.R
\name{WISEsummary}
\alias{WISEsummary}
\title{WISEsummmary}
\usage{
WISEsummary(data, DV, betweenvars = NULL, withinvars = NULL, idvar = NULL,
  CI_width = 0.95)
}
\arguments{
\item{data}{A data frame}

\item{DV}{Character vector giving the dependent variable}

\item{betweenvars}{Character vector giving the between subject variables}

\item{withinvars}{Character vector giving the within subject variables}

\item{idvar}{Character vector giving the name of the column holding subject
identifiers}

\item{CI_width}{Numeric vector giving the confidence level for computing the
confidence interval boundaries. Must be between 0 and 1, non-inclusive.}
}
\value{
A data frame
}
\description{
Within-subject Error Summary
}
\examples{
library(whoppeR)
data_summary <- WISEsummary(MemoryDrugs, DV = "time", idvar = "subject",
                            withinvars = c("shape","color"))

}

